Adding the data_source class
I realized it might be better to separate data input from the metis class. This should give the possibility to add machine specific data source accessing the data for the MetisModel class without the need of rewriting it. For example the data_source class for compass will be reading the metis simulation results from the CDB database and the class will be in the cherab.compass package.